Refactor to Ensure there is no duplicate entry of the same person in particular address book
